Created separate files for steps of creating VM groups
